+++ app/src/main/java/ua/romanik/vladislav/picsumphotos/domain/repository/IPhotosRepository.kt
@@ -4,7 +4,7 @@ import ua.romanik.vladislav.picsumphotos.domain.model.Photo
 
 interface IPhotosRepository {
 
-    suspend fun fetchPhotos(page: Int, limit: Int): List<Photo>
+    suspend fun fetchPhotos(page: Long, limit: Int): List<Photo>
 
     suspend fun fetchPhotoById(id: Long): Photo
 }
